Escape all characters sent to be executed in a shell
Because rusrv_ssh executed on the backend along the lines of
'rudial +exec/shell args', if the args part contains special
characters (e.g., $, comma, space, backtick) this messes
things up because it tries to execute in the space of
rudial rather than the shell of +exec/shell. So, we are
trying to escape everything at the rudial level which can
then resolve properly at the +exec/shell level. It is not
clear if this really works and what might happen in shells
other than bash and co. The requirement is that \x always
escapes x.