(RUSSNG-340) Move russ_deadline argument to index 0 or 1 as appropriate
The point is to hightlight the need for/use of a deadline argument.
When system call-like functions are involved, adding _deadline to
the function name makes it clear (along with a russ_ prefix) that
the function is different from the system call _and_ uses deadline
even if the system call takes a timeout of some sort.